Refactor FormUI
