+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/BaseAddPaymentMethodFragment.kt
@@ -28,10 +28,10 @@ import com.stripe.android.paymentsheet.model.PaymentSelection
 import com.stripe.android.paymentsheet.model.SupportedPaymentMethod
 import com.stripe.android.paymentsheet.paymentdatacollection.ComposeFormDataCollectionFragment
 import com.stripe.android.paymentsheet.paymentdatacollection.FormFragmentArguments
-import com.stripe.android.paymentsheet.paymentdatacollection.TransformToPaymentMethodCreateParams
 import com.stripe.android.paymentsheet.ui.AnimationConstants
 import com.stripe.android.paymentsheet.viewmodels.BaseSheetViewModel
 import com.stripe.android.ui.core.Amount
+import com.stripe.android.ui.core.FieldValuesToParamsMapConverter
 import com.stripe.android.ui.core.elements.IdentifierSpec
 import kotlinx.coroutines.launch
 
@@ -223,21 +223,18 @@ internal abstract class BaseAddPaymentMethodFragment : Fragment() {
         childFragmentManager.findFragmentById(R.id.payment_method_fragment_container)
 
     companion object {
-        private val transformToPaymentMethodCreateParams = TransformToPaymentMethodCreateParams()
 
         @VisibleForTesting
         internal fun transformToPaymentSelection(
             formFieldValues: FormFieldValues?,
             selectedPaymentMethodResources: SupportedPaymentMethod,
         ) = formFieldValues?.let {
-            transformToPaymentMethodCreateParams.transform(
-                formFieldValues.copy(
-                    fieldValuePairs = it.fieldValuePairs
-                        .filterNot { entry ->
-                            entry.key == IdentifierSpec.SaveForFutureUse ||
-                                entry.key == IdentifierSpec.CardBrand
-                        }
-                ),
+            FieldValuesToParamsMapConverter.transformToPaymentMethodCreateParams(
+                formFieldValues.fieldValuePairs
+                    .filterNot { entry ->
+                        entry.key == IdentifierSpec.SaveForFutureUse ||
+                            entry.key == IdentifierSpec.CardBrand
+                    },
                 selectedPaymentMethodResources.type
             ).run {
                 if (selectedPaymentMethodResources.type == PaymentMethod.Type.Card) {
